+++ mcxlab/examples/demo_mcxlab_basic.m
@@ -7,7 +7,7 @@
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 
 cfg.nphoton=1e7;
-cfg.vol=uint8(ones(60,60,60));
+cfg.vol=uint16(ones(60,60,60));
 cfg.srcpos=[30 30 1];
 cfg.srcdir=[0 0 1];
 cfg.gpuid=1;
